added new mobs and effects
Fixed a critical error where the spawn timer had a race condition causing mobs to stop spawning after a few waves. Fixed pacing between the change in the wave list and the spawning of mobs from that wave.  Added shaders for effects use. Note - this patch requires crediting thedevelopers.tech for the shaders.